Create LeetCode_459.py
Given a non-empty string check if it can be constructed by taking a substring of it and 
appending multiple copies of the substring together. You may assume the given string 
consists of lowercase English letters only and its length will not exceed 10000.

Example 1:

Input: "abab"
Output: True
Explanation: It's the substring "ab" twice.
Example 2:

Input: "aba"
Output: False
Example 3:

Input: "abcabcabcabc"
Output: True
Explanation: It's the substring "abc" four times. (And the substring "abcabc" twice.)

My Explanation:
==========
Hidden case: Empty string and string of length.

I am unable to understand why string of length one should give false. 
I included that corner check when my one-line solution 

"return len(s) % len(set(s)) == 0"

failed for s = "a"

Then, I thought and wrote the following solution. 

Since the original string is formed of one of its substrings, 
it means that the length of the substring MUST be a factor of
the length of the original string.

Therefore, the loop need not run more than half, because
any substring with length more than half of the length
of the original string cannot be replicated to generate the string 
because in that case the length of the pattern will be more than
the length of the original string. 

So, we start the loop from 1 indicating the length of 1. 
If it is a factor of length of the string, then we
check whether the substring of that length replicated len(s)//len(substring)
times is same as the original string. If yes, return True.
Once we are out of the loop without break, simply return False.

TC = O(n)